+++ src/components/tabs/tabs-interface.ts
@@ -3,4 +3,5 @@ export type WcsTabsAlignment = 'start' | 'center' | 'end';
 export interface WcsTabChangeEvent {
     tabName: string;
     tabIndex: number;
+    selectedKey: any;
 }